pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = ""
authors = [
    {name = "flesm",email = "ostrovskijan659@gmail.com"}
]
readme = "README.md"

[tool.poetry]
name = "auth"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { version = ">=0.115.1,<0.116.0", extras = ["standard"] }
sqlalchemy = "^2.0.41"
alembic = "^1.16.1"
python-dotenv = "^1.1.0"
asyncpg = "^0.30.0"
bcrypt = "4.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.5.0"
fastapi-mail = "^1.5.0"
pydantic-settings = "^2.9.1"
dependency-injector = "^4.48.0"

[tool.poetry.group.dev.dependencies]
black = "^24.1.0"
isort = "^5.12.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
import-linter = "^2.3"
pytest-alembic = "^0.12.1"
faker = "^37.4.0"
pytest-asyncio = "^1.0.0"
pytest-cov = "^6.2.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3

[tool.pytest.ini_options]
asyncio_mode = "auto"